some refactoring after review
