+++ dataland-frontend/tests/e2e/specs/eu-taxonomy-financials/DataIntegrity.ts
@@ -1,6 +1,7 @@
 import { describeIf } from "../../support/TestUtility";
 import { createCompanyAndGetId } from "../../utils/CompanyUpload";
-import { submitEuTaxonomyFinancialsUploadFormAndGetDataId } from "../../utils/EuTaxonomyFinancialsUpload";
+import { submitEuTaxonomyFinancialsUploadForm } from "../../utils/EuTaxonomyFinancialsUpload";
+import { generateEuTaxonomyUpload } from "../../utils/EuTaxonomyFinancialsUpload";
 import {
   CompanyInformation,
   EuTaxonomyDataForFinancials,
@@ -41,57 +42,6 @@ describeIf(
       });
     }
 
-    function fillField(divName: string, inputName: string, value?: any) {
-      if (value !== undefined) {
-        const input = value.toString();
-        if (divName === "") {
-          cy.get(`input[name="${inputName}"]`).type(input);
-        } else {
-          cy.get(`div[name="${divName}"]`).find(`input[name="${inputName}"]`).type(input);
-        }
-      }
-    }
-
-    function generateEuTaxonomyUpload(data: EuTaxonomyDataForFinancials) {
-      cy.get("select[name=financialServicesTypes]").select(data.financialServicesTypes);
-      cy.get("select[name=attestation]").select(data.attestation.toString());
-      cy.get(`input[name="reportingObligation"][value=${data.reportingObligation.toString()}]`).check();
-      if (data.eligibilityKpis !== undefined) {
-        if (data.eligibilityKpis.CreditInstitution !== undefined) {
-          fillEligibilityKpis("CreditInstitution", data.eligibilityKpis.CreditInstitution);
-        }
-        if (data.eligibilityKpis.InsuranceOrReinsurance !== undefined) {
-          fillEligibilityKpis("InsuranceOrReinsurance", data.eligibilityKpis.InsuranceOrReinsurance);
-        }
-        if (data.eligibilityKpis.AssetManagement !== undefined) {
-          fillEligibilityKpis("AssetManagement", data.eligibilityKpis.AssetManagement);
-        }
-      }
-      if (data.insuranceKpis !== undefined) {
-        fillField(
-          "",
-          "taxonomyEligibleNonLifeInsuranceActivities",
-          data.insuranceKpis.taxonomyEligibleNonLifeInsuranceActivities
-        );
-      }
-      if (data.creditInstitutionKpis !== undefined) {
-        fillField(
-          "",
-          "tradingPortfolioAndInterbankLoans",
-          data.creditInstitutionKpis.tradingPortfolioAndInterbankLoans
-        );
-        fillField("", "tradingPortfolio", data.creditInstitutionKpis.tradingPortfolio);
-        fillField("", "interbankLoans", data.creditInstitutionKpis.interbankLoans);
-      }
-    }
-
-    function fillEligibilityKpis(divName: string, data: EligibilityKpis) {
-      fillField(divName, "taxonomyEligibleActivity", data.taxonomyEligibleActivity);
-      fillField(divName, "derivatives", data.derivatives);
-      fillField(divName, "banksAndIssuers", data.banksAndIssuers);
-      fillField(divName, "investmentNonNfrd", data.investmentNonNfrd);
-    }
-
     function uploadDataAndVisitCompanyPage(
       companyInformation: CompanyInformation,
       testData: EuTaxonomyDataForFinancials
@@ -99,7 +49,7 @@ describeIf(
       createCompanyAndGetId(companyInformation.companyName).then((companyId) => {
         cy.visitAndCheckAppMount(`/companies/${companyId}/frameworks/eutaxonomy-financials/upload`);
         generateEuTaxonomyUpload(testData);
-        submitEuTaxonomyFinancialsUploadFormAndGetDataId();
+        submitEuTaxonomyFinancialsUploadForm();
         cy.visitAndCheckAppMount(`/companies/${companyId}/frameworks/eutaxonomy-financials`);
       });
     }
