+++ dataland-frontend/tests/e2e/utils/EuTaxonomyFinancialsUpload.ts
@@ -1,31 +1,69 @@
-export function fillEuTaxonomyFinancialsUploadFields(): void {
-  cy.get("select[name=financialServicesTypes]").select("Credit Institution");
-  cy.get("select[name=attestation]").select("Limited Assurance");
-  cy.get('input[id="reportingObligation-option-yes"][value=Yes]').check({ force: true });
-  cy.get('div[name="CreditInstitution"]').find("input[name=taxonomyEligibleActivity]").type("0.5", { force: true });
-  cy.get('div[name="CreditInstitution"]').find("input[name=derivatives]").type("0.1", { force: true });
-  cy.get('div[name="CreditInstitution"]').find("input[name=banksAndIssuers]").type("0.3", { force: true });
-  cy.get('div[name="CreditInstitution"]').find("input[name=investmentNonNfrd]").type("0.25", { force: true });
-  cy.get("input[name=tradingPortfolio]").type("0.6", { force: true });
-  cy.get("input[name=interbankLoans]").type("0.4", { force: true });
-}
+import {
+  EuTaxonomyDataForFinancials,
+  EligibilityKpis,
+} from "../../../../build/clients/backend/org/dataland/datalandfrontend/openApiClient/model";
 
-export function submitEuTaxonomyFinancialsUploadFormAndGetDataId(): Cypress.Chainable<string> {
+export function submitEuTaxonomyFinancialsUploadForm(): Cypress.Chainable {
   cy.intercept("**/api/data/eutaxonomy-financials").as("postCompanyAssociatedData");
   cy.get('button[name="postEUData"]').click();
-  return cy
-    .wait("@postCompanyAssociatedData")
-    .get("body")
-    .should("contain", "success")
-    .get("span[title=dataId]")
-    .then<string>(($dataId) => {
-      const id = $dataId.text();
-      return id;
-    });
+  return cy.wait("@postCompanyAssociatedData").get("body").should("contain", "success");
 }
 
-export function uploadEuTaxonomyDataForFinancials(companyId: string): Cypress.Chainable<string> {
+export function uploadDummyEuTaxonomyDataForFinancials(companyId: string): Cypress.Chainable {
   cy.visitAndCheckAppMount(`/companies/${companyId}/frameworks/eutaxonomy-financials/upload`);
-  fillEuTaxonomyFinancialsUploadFields();
-  return submitEuTaxonomyFinancialsUploadFormAndGetDataId();
+  fillEuTaxonomyFinancialsDummyUploadFields();
+  return submitEuTaxonomyFinancialsUploadForm();
+}
+
+export function generateEuTaxonomyUpload(data: EuTaxonomyDataForFinancials) {
+  cy.get("select[name=financialServicesTypes]").select(data.financialServicesTypes);
+  cy.get("select[name=attestation]").select(data.attestation.toString());
+  cy.get(`input[name="reportingObligation"][value=${data.reportingObligation.toString()}]`).check();
+  if (data.eligibilityKpis !== undefined) {
+    if (data.eligibilityKpis.CreditInstitution !== undefined) {
+      fillEligibilityKpis("CreditInstitution", data.eligibilityKpis.CreditInstitution);
+    }
+    if (data.eligibilityKpis.InsuranceOrReinsurance !== undefined) {
+      fillEligibilityKpis("InsuranceOrReinsurance", data.eligibilityKpis.InsuranceOrReinsurance);
+    }
+    if (data.eligibilityKpis.AssetManagement !== undefined) {
+      fillEligibilityKpis("AssetManagement", data.eligibilityKpis.AssetManagement);
+    }
+  }
+  if (data.insuranceKpis !== undefined) {
+    fillField(
+      "",
+      "taxonomyEligibleNonLifeInsuranceActivities",
+      data.insuranceKpis.taxonomyEligibleNonLifeInsuranceActivities
+    );
+  }
+  if (data.creditInstitutionKpis !== undefined) {
+    fillField("", "tradingPortfolioAndInterbankLoans", data.creditInstitutionKpis.tradingPortfolioAndInterbankLoans);
+    fillField("", "tradingPortfolio", data.creditInstitutionKpis.tradingPortfolio);
+    fillField("", "interbankLoans", data.creditInstitutionKpis.interbankLoans);
+  }
+}
+
+function fillEligibilityKpis(divName: string, data: EligibilityKpis) {
+  fillField(divName, "taxonomyEligibleActivity", data.taxonomyEligibleActivity);
+  fillField(divName, "derivatives", data.derivatives);
+  fillField(divName, "banksAndIssuers", data.banksAndIssuers);
+  fillField(divName, "investmentNonNfrd", data.investmentNonNfrd);
+}
+
+function fillField(divName: string, inputName: string, value?: any) {
+  if (value !== undefined) {
+    const input = value.toString();
+    if (divName === "") {
+      cy.get(`input[name="${inputName}"]`).type(input);
+    } else {
+      cy.get(`div[name="${divName}"]`).find(`input[name="${inputName}"]`).type(input);
+    }
+  }
+}
+
+function fillEuTaxonomyFinancialsDummyUploadFields(): void {
+  cy.get("select[name=financialServicesTypes]").select("Credit Institution");
+  cy.get("select[name=attestation]").select("Limited Assurance");
+  cy.get('input[name="reportingObligation"][value=Yes]').check();
 }
